Handle race when many consumers call ResetOffsets
There can be a race condition when many consumers in a group are trying
to reset offsets. This happens in the gap between checking if a zk node
exists and deleting the node.

A way to handle this is to check if the error returned by Delete is
zk.ErrNoNode and ignoring it.